Blue-green deployment
The default concurrency for deployment on multiple nodes is `serial`: one at a time.

For 2 nodes, `blue-green` is identical, but detaches every node from the load balancer before deploying to it; so in case of any failure the node stays detached and deploy stop; requests are not served by a node being deployed to, which may be doing maintenance operations.

If there were e.g. 6 nodes, `blue-green` would deploy first to 3 of them
and then to the other 3.